requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
jenkinsapi==0.3.13
multi_key_dict==2.0.3
pbr==6.1.0
python-jenkins==1.8.2
pytz==2024.2
requests==2.32.3
six==1.17.0
urllib3==2.3.0
